Fix flaky tests + add __str__ of ConsentForest (#1038)
We experienced rare test failures in CI which were reproducible by
reusing the pytest-randomly seed. In order to diagnose the cause, a
`__str__` method was added to the ConsentForest class to stringify its
primary elements (consent nodes and consent trees).

The node and tree classes already define `__str__`, and
`ConsentTree.__str__` is formatted and indented data in some cases.
So the container's string form is also newline separated and uses
`textwrap.indent` for nice presentation.

Inspection of the stringified consent data in a failing case revealed
non-unique node IDs in the test data. Tracking down the production of
these IDs, we find a `random.randint()` usage which *can* collide in
rare cases. To rectify, simply replace the fully random data with the
integer form of UUIDs -- this guarantees uniqueness with minimal code
complexity cost.