Refine and fix the container types in GlobusApp (#1035)
For user-input interfaces, we should generally declare types which
express our constraints vis-a-vis mutability of container types.
That is, if we can accept a `Mapping[K, V]`, we should declare that as
our input constraint, rather than `dict[K, V]` or
`MutableMapping[K, V]`.

The reason for this is that the variance of `K` and `V` over an
immutable mapping is more permissive. Immutable container types tend
to be covariant, whereas mutable containers are invariant.
Invariance can incorrectly penalize users by being overly strict,
rejecting subtypes of `K` and `V`.

Therefore, fix all of the input points for GlobusApp which
unnecessarily use `dict` to use `MutableMapping` instead.

Additionally, make `__init__` methods consistent in declaring a return
type of `None`. `mypy` is now lenient, allowing omission of the return
type on `__init__`, but it considers a no-arg unannotated init to be
untyped. This can lead to surprising behaviors -- for consistency,
`__init__` should always be annotated with a return type.

A typing regression test is included for the usage which first raised
this issue.