Make `GlobusApp.token_storage` public (#1040)
In service of this, the docstrings for GlobusApp needed to be fixed.
The `__init__` docstring is not used in our sphinx configuration, and
is therefore moved into the class docstring, where it will be properly
loaded and used for parameter doc. This allows the renamed attr to be
documented as an instance var.