Make ValidationError public and move 'check_uuid'
All of the changes made here, which are numerous in terms of quantity,
but all small in terms of impact, are done in service of two moves to
make the `_guards.validators` behave better.

1. ValidationError is moved to be a public interface of the SDK,
   provided by `globus_sdk.exc` alongside `GlobusSDKUsageError`

2. `utils.check_uuid` is moved to `_guards.validators.uuidlike`

(1) Is a public API change. Because we are raising a specific error
class in certain cases, it is reasonable to suppose that a caller may
want to catch that class -- in fact, we do so ourselves within the
GARE code. Users cannot do this if we do not export the symbol for the
error class.

(2) Is a purely internal change, so that we have greater uniformity
within the SDK codebase. `check_uuid` was needed in the stable SDK
modules, but its purpose is clearly the same as the other validators.
Moving things around includes changing the interface to match the
other validators (takes name first, returns the value).

New tests are added for the guards, which were previously untested.