Improve typing-time lazy import behaviors to expose attribute errors to type checkers (#1052)
* Run "tox -m freezedeps"

* Add a typing test for getattr-based lazy imports

Ensure that a missing name gets a type-error, meaning that the getattr
importer hasn't obscured other errors.

* Fix the lazy-importer at typing-time

Move the definitions of `__getattr__` and `__dir__` into the `else`
block of the `if TYPE_CHECKING` directive, meaning that they are not
evaluated at typing time.

Doing so forces `mypy` to evaluate the listed imports as though they
were the only things in the module, and to treat them and the
`__all__` list as a canonical listing of top-level `globus_sdk`
attributes. As a result, previously undetectable errors (such as a bad
module-attribute name) are properly detected.

Without this characteristic, `mypy` sees the `__getattr__` method and
assumes that any module attribute access will be valid.