Add response subtypes for different oauth2 grants (#1051)
* Add response subtypes for different oauth2 grants

The initial goal of this work is to allow a function handling a token
response, e.g., in a tokenstorage interface, to determine the context
in which the original call was made. It is now possible to check

    if isinstance(response, RefreshTokenResponse): ...

in a generic token handler.

Additionally, this opens the way for distinct token response objects
to implement methods which vary based on the known properties of their
respective grant types. For example, RefreshTokenResponse could
raise a more informative error when `decode_id_token` is called and
no `id_token` is present.

* Update src/globus_sdk/authorizers/client_credentials.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Update src/globus_sdk/authorizers/client_credentials.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Minor docstring fixes

* Update src/globus_sdk/authorizers/refresh_token.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Minor doc ordering fix

* Rename the new OAuth token response classes

Per discussion, this is a slight rename to restore the `OAuth` prefix
and reduce the suffix to `Response`.

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>