Improve control over `identity_id` extraction in experimental TokenStorage (#1055)
* Control determination of `identity_id` by context

Depending on the response context (the type of OAuthTokenResponse in
use) and the calling context (ValidatingTokenStorage or the
base class), there are various optimal behaviors for `identity_id`
determination.

In order to avoid declaring any new public interface which would need
support, this is defined as an internal-only hook,
`TokenStorage._extract_identity_id`, and given various behaviors in
different scenarios.

In the typical case, it acts exactly as it used to. The following
scenarios modify its behavior:
- on dependent token responses, it returns None
- in the ValidatingTokenStorage, it returns
  `ValidatingTokenStorage.identity_id` IFF the response is an
  OAuthRefreshTokenResponse

* Test tokenstorage handling of various types

Define tests which ensure that all of the standard TokenStorage
classes handle the variants of OAuthTokenResponse gracefully.
Refresh and dependent token responses should be accepted, in spite of
their data shapes not including `id_token`.

In the case of refresh token data, `id_token` is absent, and in the
case of dependent token data, the top-level structure is an array.

As a control against these two tests, check the behavior of an
authorization code grant response containing an `id_token`.

* Add a test for ValidatingTokenStorage on refresh

A dedicated test case stores token data twice, using a refresh token
response for the second store. It verifies that under such a usage,
the original identity information is preserved.

* Add changelog for id_token handling fixes

* Fix type checking in TokenStorage