Implement simple id-token decoder
This establishes a new helper class, `IDTokenDecoder` with a default
implementation, `DefaultIDTokenDecoder`, which implement `id_token`
decoding.

In addition to the basic decoding step, the default implementation
doubles as a cache, allowing the results of OIDC config fetching and
JWK fetching to be cached in memory, bound to the lifetime of a
well-scoped object.

The default JWT leeway of 5 minutes is now declared as a classvar,
`IDTokenDecoer.DEFAULT_JWT_LEEWAY`.