Rewrite ValidatingTokenStorage to be more modular (#1061)
* Rewrite ValidatingTokenStorage to be more modular

- Split ValidatingTokenStorage into different components which segment
  the functionality. Primarily this introduces "validators" as a
  separate piece, but it also moves identity_id reading into a very
  small dedicated class for easy code reuse.

- Two initial validator implementations have been added, each handling
  a different (stateful) part of the ValidatingTokenStorage. One
  covers ScopeRequirements, ensuring that they are met via the
  Consents API, and the other handles Identity ID, requiring that the
  value does not change once it has been set initially.

- GlobusApp is updated to construct a ValidatingTokenStorage which
  uses a ScopeRequirementsValidator and an
  UnchangingIdentityIDValidator internally, by default.

- AuthorizerFactory classes are updated to inject their own validators
  into the ValidatingTokenStorage on initialization.

- ClientCredentialsAuthorizerFactory is updated to take
  scope_requirements on init, rather than trying to pull them from the
  token storage (where they are no longer available).

* Fix IdentityIDReader and unit tests

IdentityIDReader is here updated to correctly store any value it reads
for `identity_id` -- this fixes a flaw in which the `identity_id`
attribute is persistently `None`.

Additionally, adjust several unit tests of the ValidatingTokenStorage
to respect its updated initialization path.

* Replace a unit test for _consent_client

This test was checking that we have a non-null _consent_client
attribute on ValidatingTokenStorage. That attribute no longer exists.

A better and more appropriate test for the default initialization
path is that we implicitly register the `openid` scope.

* Replace mock ValidatingTokenStorage in tests

Rather than trying to establish a mock, and thereby failing to supply
the requisite hooking interface, simply use a memory-backed storage
for unit tests of the authorizer factories.

* Further refactor ValidatingTokenStorage

Move the implementation into a nested subpackage and introduce a
TokenValidationContext object which allows the storage to pass state
down to the validators in a structured manner.

Validators are now declared as a protocol (and there is only one such
protocol), and both before_store and after_retrieve are coerced into
the same data-shape for simplicity. This also allows for batched
execution of the after_retrieve validators, when getting all token
data from storage.

Names and export symbols are updated as appropriate, and tests require
only minimal tweaks. All of the relevant elements of this system are
now provided as `globus_app`-level exports.

Internal validators are declared to encode the
AccessTokenAuthorizerFactory and RefreshTokenAuthorizerFactory checks,
but the structure by which these are applied is not yet changed.

* Refine the redesign of ValidatingTokenStorage

This is a further refinement from prior changes, which makes the
ValidatingTokenStorage initialization in GlobusApp more
straightforward.

1. Allow for `self._authorizer_factory = None` in various GlobusApp
   codepaths. This means that during initialization, functions which
   use `self._authorizer_factory` can -- at least in theory -- be
   called without error. The failure modes for this being None are
   explicitly declared as NotImplementedError in codepaths which
   should be unreachable.

2. Improve the interface for defining token data validators. They are
   now expected to be objects implementing one or both of
   `after_retrieve` and `before_store`. (This requires definition as a
   union type.)

3. Now that `_authorizer_factory: None` is allowed, the
   ValidatingTokenStorage can be initialized before
   _authorizer_factory. This enables UserApp to take responsibility
   for setting the appropriate validators for its relevant
   AuthorizerFactory types during init.

4. Default initialization of a ValidatingTokenStorage is now defined
   in a dedicated helper function,
   `build_default_validating_token_storage`.

* Simplify ValidatingTokenStorage per review

- Use a bool rather than nullability to detect whether or not
  _authorizer_factory is fully initialized

- Replace Protocol usages with an ABC

- Move the _build_default_validating_token_storage helper and mark it
  clearly as private

* Adjust tests to cover separation of validators

Validator behavior is no longer baked into authorizer factories.
Testing these usages therefore can either be done by parts or by
explicitly reconstructing realistic modalities for the authorizer
factory and validator state.

* Add more unit tests for ValidatingTokenStorage

* Add changelog for ValidatingTokenStorage changes

* Update src/globus_sdk/experimental/globus_app/validating_token_storage/validators.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Replace _authorizer_factory_initialized with stub

Rather than an initialization bool, this state can be managed by means
of a stub object which implements no-op and error behaviors. As a
result, codepaths outside of initialization don't need any awareness
of the partially-initialized state.

* Revert 58682e37, avoid a class attr for init bool

Revert 58682e37.
Also, move the bool for initialization from being a class attr to an
instance attr created at the start of init.

* Apply doc feedback from review

Co-authored-by: Kurt McKee <39996+kurtmckee@users.noreply.github.com>

* Fix import after rebase

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>
Co-authored-by: Kurt McKee <39996+kurtmckee@users.noreply.github.com>