Add a pre-build scriv step to RTD builds (#1080)
* Add a pre-build scriv step to RTD builds

When running a readthedocs PR build, automatically run `scriv collect`
if there are changelog fragments, with a `-dev` suffixed version
number.
In support of this, install `scriv` into docs environments, so that it
can be run by readthedocs. (Requires a run of `tox r -m freezedeps`.)

The end result is that PR builds will be able to see a preview of the
changelog. It won't necessarily be "fully cleaned up" in cases where
we want to make some edits at the time of the release, but it will be
close and in many cases perfect.

* Refine RTD PR changelogs

- use the PR number rather than 'dev'
- also, turn on verbose scriv logging, in case this ever fails