Improve consistency of warning style in tests
In the module where a `from globus_sdk import ...` is used,
`RemovedInV4Warning` is imported there, inlining it with other imports
and matching the documented usage for users.

In the module where `import globus_sdk` is used, the warning is
accessed as `globus_sdk.RemovedInV4Warning`. Again, this matches
user-facing documentation.

The two modules are not made consistent *with one another*, but they
are now each internally consistent.

---

Separately, fix a small issue which crept in, in which multiple tests
were written in a single test function and I therefore failed to
attach the requisite context manager.