Update RTD config to use 'uv' for faster builds
Using `build.commands`, we completely override the build process for
RTD. Per their documentation (linked in a comment), this is the
way to convert a doc build to use `uv`.

The install is done using the `uv pip` interface, which has a few
advantages:
- it can be treated as a plug-in replacement for `pip ...`
- it supports `-r` options (so we don't need to change how we're
  locking dependencies here)
- it ensures that we don't see any surprising new semantics from the
  uv lockfile (which could be hard to troubleshoot locally)

Because we're now using `build.commands`, there's no point in keeping
`jobs.pre_build` as separate config. Simply inline the relevant
command into our overridden process.