clean up scriv snippet and improve annotation
Changes implemented per review from @sirosen.

* Uncommented out header lines for scriv and
removed lines not pertinent to change.

* Updated annotation for filter_entity_type to use
Literal[] with list of five recognized entity
types as the API only accepts a single entity
type as a value.

* Updated documentation for filter_entity_type
be more explicit that it filters on only one
type. That is, you can't combine them into a
comma-delimited list.