Minor tox.ini improvement: use `base = mypy`
Declaring a `base` environment allows you to declare a parallel
environment to an existing one with some settings overridden. This
lets us avoid the generative environment listing syntax (slightly
obscure as a feature, not everyone knows about it) and having two
`commands` lists fully differentiated by a factor.

The config here is 100% equivalent to what we had before, but is
easier to read and understand for developers less immersed in tox.