Improve `RegisteredResponse` init doc & signature (#1126)
* Improve `RegisteredResponse` init doc & signature

`RegisteredResponse.__init__` allowed any keyword arguments via
`**kwargs` usage, but would later pass these to `responses.Response`
via the `add()` or `replace()` methods, at which point there would be
a `TypeError`.

This can make it hard to understand what features are being exposed
during development. e.g., "Was the parameter name `status` or
`status_code`?"

To resolve and provide an improved experience for development with
`RegisteredResponse`, the init parameters are better documented and
the signature is improved to explicitly enumerate supported args.

1. Add `:param ...:` docstring items
2. Explicitly enumerate parameters, removing `**kwargs: Any`
3. Document which parameters come from `responses.Response`, which
   come from `responses.BaseResponse`, and which ones are omitted (and
   why)
4. Update `service` and `method` to use Literals to show the values

In terms of lost functionality, technically the following three
parameters are removed in this change:

    - auto_calculate_content_length
    - passthrough
    - match_querystring

However, given that `_testing` has a weaker backwards compatibility
contract than the rest of the SDK and there are no known usages for
these, this seems like a good change rather than a bad one. We can
always add `auto_calculate_content_length` in the future. The other
two would require stronger justification, since
- match_querystring is deprecated in `responses`
- passthrough doesn't seem applicable to any normal uses of
  `globus_sdk._testing`
- we don't forbid anyone from using `responses` directly

* Update src/globus_sdk/_testing/models.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Enforce `RegisteredResponse.method` correctness

Compare in a test against the options supported by `http.HTTPMethod`,
and thereby compare against the strings supported by `responses`.

Add `OPTIONS`, `CONNECT`, and `TRACE` for completeness.

* Make `RegisteredResponse.match` default to None

In order to ensure that we aren't overriding a potentially new default
in `responses` by passing `()` explicitly to it, simply default to
`None` and only pass a value if it is non-null.

Also refine docstrings.

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>