Integrate `IDTokenDecoder` into GlobusApp
A GlobusAppConfig may now define an `IDTokenDecoder` which will be
passed to the app on init. If no decoder is defined, a default one
will be built for the user.

Because abstract base classes can't add new requirements without
breaking 3rd party implementations, token storage cannot require a
decoder. It therefore must be optional.
The decoder is attached to the underlying token storage as a new
attribute, and the attribute is runtime-checked to see if it should be
used for decoding. The attribute is wrapped in a property simply to
provide better documentation facilities (property getter docstring).

The `IDTokenDecoder` found in a GlobusApp will always be attached to
the underlying token storage. Either the default one built on init, or
the one which was passed through `GlobusAppConfig`.

The end result is that in typical GlobusApp usage:
- a DefaultIDTokenDecoder is built on init
- that decoder is used each time a token is being checked by storage
- the decoder caches `openid_configuration` and `jwk` for the lifetime
  of the GlobusApp, making us more efficient in calling out to Globus
  Auth