Add IDTokenDecoderProvider in GlobusAppConfig
This is a new protocol mirroring our other protocol definitions for
the `for_globus_app` inversion of control which allows the various
classes to define their own build hooks for globus app construction.
This simplifies the default behavior for `GlobusAppConfig` to make the
default the `DefaultIDTokenDecoder` class itself (because the class is
a provider).

For interface uniformity, this accepts the app name and the config
object, even though these are not used.