Refine IDTokenDecoder class structure
Refactor so that `JWTDecoder` is the base (abstract) class and the
implementation gets the `IDTokenDecoder` name. This fits so long as
the abstract base doesn't work in terms of OIDC id_tokens but instead
of generic JWTs instead.

To fit this phrasing, and still provide control via subclassing to
jwt.decode parameters, all params passed to the `decode()` call other
than the token are removed. Instead, these are attributes of the
decoder.

This allows for a tighter definition of the abstract `decode()`
method, taking only a token string as an input.

Tests, docs, and the changelog are all updated to fit the new naming.