Improve the initialization ordering of GlobusApp
We have a cyclic dependency in GlobusApp init, in which we want a
self-registered client for `openid` scoped consent API usage, but we
need this as an input to various constructs before we are ready to
register any clients in the app. We have a "cycle breaker" bool flag
in app init which gets used to work around this behavior.

We cannot break the cycle fully in the current design: apps require
an auth client, and clients support use with apps.
But we can rephrase the cycle to make it easier to understand and
maintain.

1.  Externalize the app-attachment behavior of a client in a new
    `BaseClient.attach_globus_app()` method. This method is used to
    implement the client init behaviors (except for a slightly subtle
    one around environment setting), and replaces
    `BaseClient._finalize_app()`.

    This allows for a client to be instantiated and -- so long as no
    app, scopes, or authorizer are set on that client -- late-bound to a
    GlobusApp.

2.  Update `GlobusApp.__init__` to create the openid-scoped auth client
    without passing `app=self` or `app_scopes`; instead, late-bind
    these parameters to the client after the app is initialized to the
    point that it can handle the client's callbacks to the app.

    This is a more straightforward cycle-breaker to read and maintain
    than the prior internal bool tracking. Semantically, it does not
    change the behavior of the app (the now-unguarded cache-clearing
    call is a no-op on a new app).

3.  As an unanticipated benefit, `BaseClient.__init__` has
    significantly improved readability because app-attachment logic is
    all consolidated (moreso than it was under `_finalize_app()`).

Net-net, we have a new method to support,
`BaseClient.attach_globus_app()`, but it has a strict contract about
being called once per client and rejecting any internal data
conflicts. It also documents itself as being called during
initialization if an app is passed.