Expose `RequestsTransport.headers` dict
The dict itself needs very little explanation -- it's an
instance-level setting which lets a user add a header to a transport
imperatively, simply by doing

    >>> client.transport.headers["My-Header"] = "foo"

or similar.

Its interactions with `User-Agent` and `X-Globus-Client-Info`,
however, are more subtle. These two values act as "dynamic headers",
meaning we either would need to split the headers into two locations
-- the static and the dynamic ones -- or else find a mechanism to keep
these data synchronized. After examining a split static/dynamic state,
it is notably harder to explain to users, harder to use, and does not
allow as much freedom to rewrite or manipulate these data. Therefore,
a new update-contract is put in place around
`RequestsTransport.headers`:

- Whenever `RequestsTransport.user_agent` is written (it is a
  property), it will write back into `headers`. This is unconditional
  and does not worry about overwriting data -- if a user is setting
  the `User-Agent` string manually, they should refrain from
  manipulating this property.

- Whenever `RequestsTransport.globus_client_info` is updated via
  `add()`, it triggers a callback to update `headers`. This mirrors
  the contract around `user_agent` being written as a property.

For almost all usages, this works fine, but there are a few cases in
the testsuite where this new callback-based mechanism results in
surprising behavior. Because tests are setting
`transport.globus_client_info.infos = []` in order to strictly control
the header, they are "desynced". In support of these test scenarios, a
new `clear()` method is added.
A review of usages reveals that no callers outside of the SDK
testsuite are relying on this behavior today.