Test `mypy` at our boundary versions (#1143)
During development, developers on lower Python versions (e.g. '3.10'
in this case) can see `mypy` errors which are not evident when `mypy`
runs on a higher supported Python version.

To ensure that type checking not only passes for all maintainers, but
also to guarantee that downstream consumption of SDK types works on
all suported Pythons, test `mypy` on our highest and lowest Python
versions explicitly. Doing so exposes the fact that `typing.Self` is
new in 3.11 and therefore not safe to use directly. Instead, we need a
`typing_extensions` dispatch.

The new testing is driven by `tox`, and by changes to CI:
- tox lists `mypy-py3.8` and `mypy-py3.13` environments, effectively
  as aliases for `mypy` (the default env_list is unchanged)
- the "Quality" job will now include a Python 3.8 interpreter
- the "Quality" job now runs `mypy-py3.8` and `mypy-py3.13`
  explicitly[^1]

[^1]: `mypy` and `mypy-py3.13` should be equivalent in this
environment, but choosing the explicit option gives less ambiguous
output in the run.