Initial tutorial rewrite for GlobusApp
This rewrite is the first step towards creating a better "Getting
Started" experience for our users without having to redirect them
*away* from the tutorial.

The new tutorial is intentionally very light on detail. It is not a
GlobusApp user-guide and it is not concerned with explaining advanced
usage scenarios or even such details as how tokenstorage is wired up
by default. That is left for future getting-started docs.

The original tutorial is (at least temporarily) retained under the
name `legacy_tutorial` so that we can continue to accommodate users
who are building or maintaining applications built in that way.
Ideally it would be removed, but it would probably (?) need
replacement with new narrative documentation which covers how you
can build applications without GlobusApp. Otherwise, users who have
existing applications will be left without the means to understand the
code they have inherited.

The legacy tutorial is hardly changed. It now has a warning at the
start and no longer links to the legacy examples.

The new tutorial copies only the client creation documentation, but
otherwise involves new steps and examples.