Improve how BaseClient.environment evaluates
Rather than erroring on mismatch with an `app` early, rely on the
check in `attach_globus_app`.

In order to do this, the behavior when setting `environment` is
changed very slightly to explicitly check and use the init parameter
if passed, and to distinguish that case from the env-var loading case.
Doing so allows us to ensure that the init parameter is higher
precedence than `app.config.environment`, but the env var is lower
precedence -- resulting in the desired error checking behavior.