Remove an overly strict test parameter matcher (#1168)
The presence of this matcher makes the fixture unusable in downstream
applications if they don't keep pace with all changes to the data
model made in `_testing`. As a result, this causes the current
globus-cli `sdkmain` builds to fail as they pick up on the new
`run_managers` and `run_monitors` fields, but cannot possibly send
those parameters because support has not been implemented yet.

Removing the matcher does not harm SDK testing and makes the default
fixture more usable by being more lenient. It is confirmed to make CLI
testing pass with a `localsdk` CLI test run.