Remove an unnecessarily strict matcher
Similar to #1168

Trying to use a mock response with payload param matching makes
development more difficult because a downstream testsuite *MUST*
provide all of the exact parameters declared or it gets no data. This
contradicts the goal of `_testing`, which is to provide easy,
low-friction mocking for downstream consumers.

If stricter matchers are a desirable feature, they can be opt-in by
making non-`default` case mocks which have these requirements.
(Essentially, the inverse of the `lenient` variant mock which is being
removed here.)