Fix typing errors and raise on unreachable types
Fix typing errors where `... | None` was passed for
`requested_scopes`.

Additionally, to handle cases in GlobusApp, raise usage errors if an
app is ever doing a login without provided scopes in the authorization
parameters. These conditions aren't unreachable based on the types
which are used, but a user who configures things to reach these
conditions is engaged in explicitly unsupported usage (e.g., trying to
give explicit parameters to a client app which don't include scopes).