Make 'GlobusAPIError.code' default to 'None'
In the source tree, the default is changed from `"Error"` to `None`.
In the tests, various direct tests are updated, as are those which
inspect the arguments passed to `super().__init__` and which feed into
the repr for an API error.

There are no paths in the source tree of the SDK itself which depend
upon the old default of `"Error"`.