Remove the 'MutableScope' type (#1198)
* Remove the 'MutableScope' type

This is deprecated in 3.x in favor of the 'Scope' type. As part of our
4.0 release, we will be removing the 'MutableScope' type.

This removes the type itself, and then handles all of the cascading
fixes this necessitates.
No update to the upgrading doc is included yet, as we have additional
plans to make changes to the scope interfaces.

* Minor fix to a mypy-test

`Scope.scopes2str` does not exist -- this test was created by a
`MutableScope` replacement but the `Scope` and `MutableScope`
interfaces differ here. Remove the bad test case.