Convert GCS payload types from "_set_opt..."
These `_set_opt...` helpers on the Payload class are mostly
unnecessary now that we have defaults of MISSING.

They are the way that nontrivial conversions are managed, so these
need to be replaced with use of helpers from `globus_sdk._remarshal`.
To aid in this, two new helpers are added to `_remarshal`.

This removes the last use of `Payload._set_optints`, and is therefore
paired with removal of that helper. Likewise,
`Payload._set_optstrlists` is also removed as it is no longer used.