Revert "Fix license deprecation warning (#1223)"
This reverts commit 0f3b9bb5be9466468d1f41ed0568aaf75ff53e9c.

When this is re-introduced, the build-system config must be updated.
Specifically, setuptools v77.0.0 or higher must be required
(currently it's set to v61.2 or higher).

Note that this will mean the package cannot be built from a `.tar.gz`
on Python 3.8 because setuptools dropped support for Python 3.8
in an earlier release.