Remove transfer_client param to data constructors
Since v0.x of the SDK (specifically, since 2e78893de4 ), we have
supported `transfer_client` as a parameter to `TransferData`, doing
the `get_submission_id()` call at initialization time.

In order to avoid a data constructor which accepts its relevant
client, and the burden this can impose (e.g., in a multi-user app,
keeping the two synced), we've been in transition to a model where the
client is responsible for making this call and writing the value back
into the payload at submission time.

The `submit_transfer` and `submit_delete` methods already support
fetching a submission ID on init, but they need a small tweak to allow
for `MISSING` values to trigger that fetch.

Docs and tests are updated, and a new section of the upgrading guide
covers how to handle the change.