Use 'is MISSING' to compare for missing-ness
Add a TYPE_CHECKING -time definition of the MissingType and MISSING
sentinel as an enum and enum member. This lets type checkers
understand that an identity comparison against the MISSING sentinel is
a valid way to type-narrow.

Update the local flake8 plugin to scan for
`isinstance(..., MissingType)` and treat that as an error.

Update all of the usages flagged by this new lint.