Add a type checking test for MISSING
This confirms that `MISSING` and `MissingType` behave appropriately at
type checking time.

Because `mypy` struggles with usages like `assert_type(x, MissingType)`
when `MissingType` is defined as a type alias, simply remove the
`_MissingEnum` placeholder and make the type of the enum `MissingType`
for the type-checking definition.
This is also a little shorter.