Make 'v1' token storage less accessible
This is a small internal change which reduces the size of the
``globus_sdk.token_storage`` top-level interface.

It has impact on code completions and the general developer experience
when using these legacy tools.

A new section is added to the upgrading guide on the changes in names
(not previously covered).