Refactor to expose `RetryConfiguration`
This change fully separates the retry config from the transport.

RetryConfiguration is a "single pane of glass" view of how a client is
configured to perform retries, clients now have a
`retry_configuration` attribute which provides all of the
retry-specific configuration. This combines details which were
previously in the transport class and details which were attached to
the retry check collection.

Because the configuration is built per-client, it can be reconfigured
by simple attribute access. For ease of use, the `tune()` context
manager was added to the transport objects, and is therefore "ported"
to the retry configuration.