Deprecate the ComputeClient class
This class (implemented as a subclass but functionally an alias) is
moved to a dedicated module to provide warn-on-import semantics.

It is now documented without autodoc, so that we can prcisely control
the content shown to users.

Tests are updated to catch and ignore the deprecation warnings.

A minor flaw in the 'ensure_expores_are_documented.py' script is
corrected, so that it will see the `.. py:class::` directive.

Because this is the first element in `__init__.pyi` with an associated
import-time deprecation warning, some new handling of warning
suppression is needed (vs prior examples).