Move 'scopes_to_scope_list' to private function
This function was supported as a public interface but only had one
known usage site, in `GlobusApp`. It has been removed and replaced
with an internal method, `GlobusApp._iter_scopes`.

Unit tests are updated to test the new replacement helper. Because the
helper has a tighter contract in terms of what it supports and in what
context, it is greatly simplified in the process as well. The older
helper structure shows evidence of past iterations in terms of
handling nested iterables.