Make ScopeParser.serialize handle the empty case
Rather than a helper specific to Globus Auth components doing this as
a post-processing step, handling for `""` as the output is now baked
into `ScopeParser.serialize`. By default, it will raise an error if it
finds that the serialized data is `""`.

For example, `ScopeParser.serialize([])` errors.

A new flag, `reject_empty=False`, can be passed to explicitly disable
the check, if users have a desire to handle an unknown iterable and
check if it's empty *after* serialization.