Rename 'set_subscription_admin_verified_id'
Remove the '_id' suffix, per discussion with the Groups team.
The new name is 'set_subscription_admin_verified'.

The old function is still available as an alias, but emits a
`RemovedInV4Warning`.

The test fixture data in `globus_sdk._testing` is slightly adjusted, but
not considered a breaking change due to the looser compatibility policy
around the provided fixtures.