Remove the RTD preview GitHub Workflow
This build now fails for us with a message about action version
pinning (presumably because we have enabled the requirement for hash
versioning).

Checking up on the action repo, it currently has a deprecation notice.

After enabling PR builds via their new integration (a GitHub App), I am
removing this build. We should continue to get PR builds, and if we do
not we can investigate without "noise" from this failing build.