Add guidance on 'base_path' to upgrading doc
Because this only impacts users who are directly using HTTP methods, the
expected impact of this change is somewhat limited. As such, it is
placed at the end of the guide.

In brief, the guide lists which clients and which methods of said
clients are impacted, as well as giving notes on:

- how the mapping of usages to URIs has changed
- impact to the testing tools
- compatible usage which does the same thing on both versions