Replace 'resource ownership' with context managers
Rather than attempting to implicitly track the client<->app association,
the SDK will allow apps and clients to track a suite of resources which
they should close when they are closed.

In the case of the client, the only resource to close is the transport
object, and only if it is not passed explicitly.

In the case of the app, the resource to close is the attached
ValidatingTokenStorage, but only if the underlying token storage was not
passed explicitly. This is slightly odd, in that the app has to choose
whether it closes `self._token_storage` or `self.token_storage`, but
there is almost no observable impact from this choice.
Closing the validating storage leaves an avenue open for that `close()`
to close other resources in the future if necessary.

Clients now support context manager usage to call `close()` on exit.